refactor(rebalancer): rename pending bridge readonly reader
