+++ src/adapter/l2Bridges/BaseL2BridgeAdapter.ts
@@ -12,7 +12,7 @@ import {
   SolanaTransaction,
   isDefined,
 } from "../../utils";
-import { PendingBridgeAdapterName, PendingBridgeRedisReader } from "../../rebalancer/utils/PendingBridgeRedis";
+import { CctpOftReadOnlyClient, PendingBridgeAdapterName } from "../../rebalancer/clients/CctpOftReadOnlyClient";
 import { TransferTokenParams } from "../utils";
 
 const DEFAULT_PENDING_WITHDRAWAL_LOOKBACK_PERIOD_SECONDS = 7200;
@@ -22,7 +22,7 @@ export abstract class BaseL2BridgeAdapter {
   protected l1Bridge: Contract;
   protected readonly hubPoolAddress: EvmAddress;
   protected readonly spokePoolAddress: Address;
-  protected pendingBridgeRedisReader?: PendingBridgeRedisReader;
+  protected pendingBridgeRedisReader?: CctpOftReadOnlyClient;
   // Whether either of these two are defined is determined at construction.
   // The solana bridge defines `svmProvider` while the EVM bridges define `l2Signer`.
   protected readonly l2Signer: Signer | undefined;
@@ -69,7 +69,7 @@ export abstract class BaseL2BridgeAdapter {
     return [];
   }
 
-  setPendingBridgeRedisReader(pendingBridgeRedisReader?: PendingBridgeRedisReader): void {
+  setPendingBridgeRedisReader(pendingBridgeRedisReader?: CctpOftReadOnlyClient): void {
     this.pendingBridgeRedisReader = pendingBridgeRedisReader;
   }
