+++ src/clients/bridges/AdapterManager.ts
@@ -14,7 +14,6 @@ import {
   isDefined,
   winston,
   Signer,
-  SVMProvider,
   TransactionResponse,
   assert,
   EvmAddress,
@@ -30,11 +29,11 @@ import { SpokePoolClient, HubPoolClient, SpokePoolManager } from "../";
 import { CHAIN_IDs, TOKEN_SYMBOLS_MAP } from "@across-protocol/constants";
 import { BaseChainAdapter } from "../../adapter";
 import { TransferTokenParams } from "../../adapter/utils";
-import { PendingBridgeRedisReader } from "../../rebalancer/utils/PendingBridgeRedis";
+import { CctpOftReadOnlyClient } from "../../rebalancer/clients/CctpOftReadOnlyClient";
 
 export class AdapterManager {
   public adapters: { [chainId: number]: BaseChainAdapter } = {};
-  protected readonly pendingBridgeRedisReader?: PendingBridgeRedisReader;
+  protected readonly pendingBridgeRedisReader?: CctpOftReadOnlyClient;
 
   // Some L2's canonical bridges send ETH, not WETH, over the canonical bridges, resulting in recipient addresses
   // receiving ETH that needs to be wrapped on the L2. This array contains the chainIds of the chains that this
@@ -51,7 +50,7 @@ export class AdapterManager {
     if (!spokePoolClients) {
       return;
     }
-    this.pendingBridgeRedisReader = new PendingBridgeRedisReader(logger);
+    this.pendingBridgeRedisReader = new CctpOftReadOnlyClient(logger);
     this.spokePoolManager = new SpokePoolManager(logger, spokePoolClients);
     const spokePoolAddresses = Object.values(this.spokePoolManager.getSpokePoolClients()).map(
       (client) => client.spokePoolAddress
@@ -84,7 +83,7 @@ export class AdapterManager {
       return Object.fromEntries(
         SUPPORTED_TOKENS[chainId]?.map((symbol) => {
           const spokePoolClient = this.spokePoolManager.getClient(chainId);
-          let l2SignerOrProvider: Signer | SVMProvider | undefined;
+          let l2SignerOrProvider;
           if (isEVMSpokePoolClient(spokePoolClient)) {
             l2SignerOrProvider = spokePoolClient.spokePool.signer;
           } else if (isSVMSpokePoolClient(spokePoolClient)) {
@@ -109,7 +108,7 @@ export class AdapterManager {
         return {};
       }
       const spokePoolClient = this.spokePoolManager.getClient(chainId);
-      let l2SignerOrSvmProvider: Signer | SVMProvider | undefined;
+      let l2SignerOrSvmProvider;
       if (isEVMSpokePoolClient(spokePoolClient)) {
         l2SignerOrSvmProvider = spokePoolClient.spokePool.signer;
       } else if (isSVMSpokePoolClient(spokePoolClient)) {
