+++ src/rebalancer/README.md
@@ -183,7 +183,7 @@ accounting) and should not initiate new rebalances.
 The read-only mode still initializes adapters (with an empty route set) so `getPendingRebalances(account)` and
 `getPendingOrders()` remain available without coupling callers to a specific operational rebalancing mode.
 
-The OFT and CCTP adapters also expose their pending bridge-pre-deposit Redis schema through `src/rebalancer/utils/PendingBridgeRedis.ts`. Adapter-side bridge accounting uses that readonly reader to ignore rebalancer-owned OFT/CCTP transfers instead of instantiating rebalancer adapters inside `AdapterManager`.
+The OFT and CCTP adapters also expose their pending bridge-pre-deposit Redis schema through `src/rebalancer/clients/CctpOftReadOnlyClient.ts`. Adapter-side bridge accounting uses that readonly reader to ignore rebalancer-owned OFT/CCTP transfers instead of instantiating rebalancer adapters inside `AdapterManager`.
 
 ### Future mode extensibility
